apps/api/pyproject.toml
[project]
name = "vm-builder-api"
version = "0.1.0"
description = "VM Builder API"
requires-python = ">=3.12"
dependencies = [
  "alembic>=1.13.0,<2.0.0",
  "fastapi>=0.115.0,<1.0.0",
  "uvicorn[standard]>=0.30.0,<1.0.0",
  "sqlalchemy>=2.0.0,<3.0.0",
  "psycopg[binary]>=3.1.0,<4.0.0",
  "pydantic-settings>=2.3.0,<3.0.0",
  "redis>=5.0.0,<6.0.0",
  "rq>=1.16.0,<2.0.0",
  "httpx>=0.27.0,<1.0.0",
  "python-jose[cryptography]>=3.3.0,<4.0.0",
  "passlib[bcrypt]>=1.7.4,<2.0.0",
  "bcrypt<4.1",
  "email-validator>=2.2.0,<3.0.0",
  "slowapi>=0.1.9,<1.0.0",
  "ldap3>=2.9.1,<3.0.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.2.0,<9.0.0",
]

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["app*"]
